Handly tested MODULE. Realized 4 types of extracting
